convert demos to TypeScript
