+++ packages/core/src/demo/status-bot.ts
@@ -17,21 +17,19 @@
  *
  */
 
-//@ts-check
-
-require('dotenv').config();
-
-const {APIClient} = require('@wireapp/api-client');
-const {ClientType} = require('@wireapp/api-client/src/client/');
-const logdown = require('logdown');
-const {Account} = require('@wireapp/core');
-const {MemoryEngine} = require('@wireapp/store-engine');
+import {APIClient} from '@wireapp/api-client';
+import {ClientType} from '@wireapp/api-client/src/client/';
+import {Account} from '@wireapp/core';
+import {MemoryEngine} from '@wireapp/store-engine';
+import * as logdown from 'logdown';
 
 const {name, version} = require('../../package.json');
 const logger = logdown('@wireapp/core/demo/StatusBot', {
   logger: console,
   markdown: false,
 });
+
+require('dotenv').config();
 
 const CONVERSATION_ARGUMENT_INDEX = 2;
 const conversations = process.argv[CONVERSATION_ARGUMENT_INDEX];
@@ -47,7 +45,7 @@ if (!message) {
   logger.warn('Message is not set. Will post a default message.');
 }
 
-['WIRE_STATUS_BOT_EMAIL', 'WIRE_STATUS_BOT_PASSWORD'].forEach((envVar, index, array) => {
+['WIRE_STATUS_BOT_EMAIL', 'WIRE_STATUS_BOT_PASSWORD'].forEach((envVar, _, array) => {
   if (!process.env[envVar]) {
     logger.error(`Error: Environment variable "${envVar}" is not set. Required variables: ${array.join(', ')}.`);
     process.exit(1);
@@ -61,18 +59,16 @@ if (!message) {
     password: process.env.WIRE_STATUS_BOT_PASSWORD,
   };
 
-  try {
-    const engine = new MemoryEngine();
-    await engine.init('');
+  const engine = new MemoryEngine();
+  await engine.init('');
 
-    const apiClient = new APIClient({store: engine, urls: APIClient.BACKEND.PRODUCTION});
-    const account = new Account(apiClient);
-    await account.login(login);
+  const apiClient = new APIClient({urls: APIClient.BACKEND.PRODUCTION});
+  const account = new Account(apiClient, () => Promise.resolve(engine));
+  await account.login(login);
 
-    const text = message || `I am posting from ${name} v${version} (Build #${process.env.TRAVIS_BUILD_NUMBER}). 🌞`;
-    const payload = account.service.conversation.createText(text).build();
-    conversationIds.forEach(async conversationId => account.service.conversation.send(conversationId, payload));
-  } catch (error) {
-    logger.error('Error:', error.stack);
+  const text = message || `I am posting from ${name} v${version}. 🌞`;
+  for (const conversationId of conversationIds) {
+    const payload = account.service.conversation.messageBuilder.createText(conversationId, text).build();
+    await account.service.conversation.send(payload);
   }
-})();
+})().catch(error => logger.error(error));
